Break Debouncing click listener into different view file.
This is so that the copyright notice is less confusing
